+++ packages/interweave-emoji-picker/tests/Emoji.test.js
@@ -27,7 +27,7 @@ describe('Emoji', () => {
   };
 
   beforeEach(() => {
-    props.emoji = EmojiData.getInstance('en').EMOJIS['🐈️'];
+    props.emoji = EmojiData.getInstance('en').EMOJIS['1F408'];
   });
 
   it('renders an emoji', () => {